Delete `ov_models` target
