+++ src/common/transformations/tests/CMakeLists.txt
@@ -16,6 +16,7 @@ ov_add_test_target(
         gmock
         func_test_utils
         openvino::offline_transformations
+        openvino::reference
         sharedTestClasses
         ov_lpt_models
     ADD_CLANG_FORMAT
