+++ src/tests/functional/plugin/shared/CMakeLists.txt
@@ -62,7 +62,6 @@ ov_add_target(
                 openvino::pugixml
                 common_test_utils
                 func_test_utils
-                ov_models
                 ov_lpt_models
                 sharedTestClasses
             PRIVATE
