+++ src/tests/functional/shared_test_classes/CMakeLists.txt
@@ -19,7 +19,6 @@ ov_add_target(
         LINK_LIBRARIES
             PRIVATE
                 func_test_utils
-                ov_models
 )
 
 ov_build_target_faster(${TARGET_NAME}
