+++ src/tests/functional/shared_test_classes/src/base/utils/generate_inputs.cpp
@@ -18,7 +18,7 @@
 #include "shared_test_classes/base/ov_subgraph.hpp"
 #include "shared_test_classes/single_op/roi_align.hpp"
 
-#include "ov_models/utils/data_utils.hpp"
+#include "common_test_utils/data_utils.hpp"
 
 namespace ov {
 namespace test {
