--- src/tests/ov_helpers/ov_models/CMakeLists.txt
@@ -1,36 +1,0 @@
-# Copyright (C) 2018-2023 Intel Corporation
-# SPDX-License-Identifier: Apache-2.0
-#
-
-set(TARGET_NAME ov_models)
-
-set(PUBLIC_HEADERS_DIR "${CMAKE_CURRENT_SOURCE_DIR}/include")
-
-ov_add_target(
-        NAME ${TARGET_NAME}
-        TYPE STATIC
-        ROOT ${PUBLIC_HEADERS_DIR}
-        INCLUDES
-            PUBLIC
-                "$<BUILD_INTERFACE:${PUBLIC_HEADERS_DIR}>"
-        ADDITIONAL_SOURCE_DIRS
-            ${CMAKE_CURRENT_SOURCE_DIR}/src
-        LINK_LIBRARIES
-            PUBLIC
-                openvino::runtime
-                openvino::reference
-                openvino::interpreter_backend
-                openvino::runtime::dev
-                common_test_utils
-        ADD_CLANG_FORMAT
-)
-
-ov_build_target_faster(${TARGET_NAME}
-    UNITY
-    PCH PRIVATE "src/precomp.hpp"
-)
-
-# install & export
-
-ov_developer_package_export_targets(TARGET ${TARGET_NAME}
-                                    INSTALL_INCLUDE_DIRECTORIES "${PUBLIC_HEADERS_DIR}/")
