+++ src/tests/test_utils/common_test_utils/include/common_test_utils/node_builders/constant.hpp
@@ -3,10 +3,10 @@
 //
 #pragma once
 
+#include "common_test_utils/data_utils.hpp"
 #include "common_test_utils/ov_tensor_utils.hpp"
 #include "openvino/core/node.hpp"
 #include "openvino/op/constant.hpp"
-#include "ov_models/utils/data_utils.hpp"
 
 namespace ov {
 namespace test {
